Update SIK_JoinLobby_AsyncFunction.cpp
Removed the Steam Result enum from the broadcasting of the delegate; we were casting from EChatRoomEnterResponse to EResult, which would give unintended consequences.
Also removed the use of ESIK_ChatRoomEnterResponse::None, as the Steam API would return ChatRoomEnterResponseError in the event an error occurred. This "None" could cause issues if a user tried to cast back to an EChatRoomEnterResponse.